+++ Streetcode/Streetcode.XUnitTest/MediatRTests/Mocks/StreetcodeRepositoryMock.cs
@@ -1,0 +1,38 @@
+﻿namespace Streetcode.XUnitTest.MediatRTests.Mocks;
+
+using Microsoft.EntityFrameworkCore.Query;
+using Moq;
+using Streetcode.DAL.Entities.Streetcode;
+using Streetcode.DAL.Repositories.Interfaces.Base;
+using System.Linq.Expressions;
+
+internal partial class RepositoryMocker
+{
+    /// <summary>
+    /// Mocks streetcode repository.
+    /// </summary>
+    /// <returns>Returns mocked repository. </returns>
+    public static Mock<IRepositoryWrapper> GetStreetcodeRepositoryMock()
+    {
+        var streetCodes = new List<StreetcodeContent>
+            {
+                new StreetcodeContent() { Id = 1, Title = "First streetcode content title" },
+                new StreetcodeContent() { Id = 2, Title = "Second streetcode content title" },
+                new StreetcodeContent() { Id = 3, Title = "Third streetcode content title" },
+                new StreetcodeContent() { Id = 4, Title = "Fourth streetcode content title" },
+            };
+
+        var mockRepo = new Mock<IRepositoryWrapper>();
+
+        mockRepo.Setup(x => x.StreetcodeRepository.GetAllAsync(It.IsAny<Expression<Func<StreetcodeContent, bool>>>(), It.IsAny<Func<IQueryable<StreetcodeContent>, IIncludableQueryable<StreetcodeContent, object>>>()))
+            .ReturnsAsync(streetCodes);
+
+        mockRepo.Setup(x => x.StreetcodeRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<StreetcodeContent, bool>>>(), It.IsAny<Func<IQueryable<StreetcodeContent>, IIncludableQueryable<StreetcodeContent, object>>>()))
+            .ReturnsAsync((Expression<Func<StreetcodeContent, bool>> predicate, Func<IQueryable<StreetcodeContent>, IIncludableQueryable<StreetcodeContent, object>> include) =>
+            {
+                return streetCodes.FirstOrDefault(predicate.Compile());
+            });
+
+        return mockRepo;
+    }
+}
